Added disk SMART monitoring to ugreen-diskiomon
Added SMART monitoring based on smartctl. The led now turns red if the corresponding disk fails (smartctl -H does not show PASSED).
Optimized the pooling loop to only iterate the existing drives at the execution of the script
The led will be turned off if no disk in installed in the slot.