Added SMART monitoring based on smartctl, and mapped the devices based on HCTL (#5)
* Added disk SMART monitoring to ugreen-diskiomon

Added SMART monitoring based on smartctl. The led now turns red if the corresponding disk fails (smartctl -H does not show PASSED).
Optimized the pooling loop to only iterate the existing drives at the execution of the script
The led will be turned off if no disk in installed in the slot.

* fixed some silly mistake

* Update ugreen-diskiomon

The sata devices are now mapped to leds based on hctl by default.
Added an option to map devices based on serial number. (serial numbers must be manually configured)
Fixed a bug in SMART polling invertal control.
Failed disk will be marked and left untouched. The failure date will be recorded.

* fix an error in ugreen_diskiomon

---------

Co-authored-by: Yuhao Zhou <miskcoo@gmail.com>